requirements.txt
apache-airflow-providers-mongo==3.0.0
cached-property==1.5.2
cachelib==0.9.0
dnspython==2.3.0
docutils==0.19
email-validator==1.3.1
et-xmlfile==1.1.0
exceptiongroup==1.1.1
Flask==2.2.3
Flask-AppBuilder==4.1.4
Flask-Babel==2.0.0
Flask-Caching==2.0.2
Flask-JWT-Extended==4.4.4
Flask-Login==0.6.2
Flask-Session==0.4.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.1.1
idna==3.4
importlib-metadata==4.13.0
importlib-resources==5.12.0
jsonschema==4.17.3
lazy-object-proxy==1.9.0
Markdown==3.4.1
markdown-it-py==2.2.0
MarkupSafe==2.1.2
marshmallow==3.19.0
marshmallow-enum==1.5.1
marshmallow-oneofschema==3.0.1
marshmallow-sqlalchemy==0.26.1
mdit-py-plugins==0.3.5
mdurl==0.1.2
multidict==6.0.4
numpy==1.21.6
openpyxl==3.1.2
packaging==23.0
pandas==1.3.5
pathspec==0.9.0
pendulum==2.1.2
psutil==5.9.4
pycparser==2.21
Pygments==2.14.0
PyJWT==2.6.0
pymongo==3.13.0
pyrsistent==0.19.3
python-daemon==3.0.1
python-dateutil==2.8.2
python-nvd3==0.15.0
python-slugify==8.0.1
pytz==2022.7.1
pytzdata==2020.1
PyYAML==6.0
requests==2.28.2
requests-toolbelt==0.10.1
spotipy==2.22.1
SQLAlchemy==1.4.46
SQLAlchemy-JSONField==1.0.1.post0
SQLAlchemy-Utils==0.40.0
sqlparse==0.4.3
urllib3==1.26.15
wrapt==1.15.0
yarl==1.8.2
zipp==3.15.0